Move YoutubeVoting to voting module
