+++ inside-java/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 
 dependencies {
     implementation(project(":publication"))
-    implementation(project(":youtube"))
+    implementation(project(":voting"))
     implementation("com.amazonaws:aws-lambda-java-events:2.2.7")
     implementation("com.amazonaws:aws-lambda-java-core:1.2.0")
     implementation("com.amazonaws:aws-lambda-java-log4j2:1.1.0")
