+++ publication/build.gradle.kts
@@ -8,7 +8,7 @@ repositories {
 }
 
 dependencies {
-    api(project(":youtube"))
+    api(project(":voting"))
     api(project(":news-queue"))
     api("org.jetbrains.kotlinx:kotlinx-coroutines-jdk8:1.3.9")
 }
