+++ voting/src/main/kotlin/by/jprof/telegram/opinions/voting/config/beans.kt
@@ -1,10 +1,14 @@
 package by.jprof.telegram.opinions.voting.config
 
 import by.jprof.telegram.opinions.voting.VotesDAO
+import by.jprof.telegram.opinions.voting.YoutubeVoting
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
 
 val votingBeans = module {
+    single {
+        YoutubeVoting(get(), get(), get(), get())
+    }
     single {
         VotesDAO(
             get(),
