+++ youtube/build.gradle.kts
@@ -8,6 +8,6 @@ repositories {
 }
 
 dependencies {
-    api(project(":voting"))
+    api(project(":components"))
     api("com.google.apis:google-api-services-youtube:v3-rev222-1.25.0")
 }
