+++ youtube/src/main/kotlin/by/jprof/telegram/opinions/youtube/config/beans.kt
@@ -1,6 +1,5 @@
 package by.jprof.telegram.opinions.youtube.config
 
-import by.jprof.telegram.opinions.youtube.YoutubeVoting
 import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport
 import com.google.api.client.json.jackson2.JacksonFactory
 import com.google.api.services.youtube.YouTube
@@ -9,10 +8,6 @@ import org.koin.core.qualifier.named
 import org.koin.dsl.module
 
 val youtubeBeansModule = module {
-    single {
-        YoutubeVoting(get(), get(), get(), get())
-    }
-
     single {
         YouTube
             .Builder(
